+++ test/misc_test.dart
@@ -6,5 +6,15 @@ void main() {
     test('sleep', () {
       expect(sleep(100), completes);
     });
+
+    test('funcTimes', () {
+      String sayHello([String? str]) {
+        return 'Hello${str ?? ''}';
+      }
+
+      expect(funcTimes(2, sayHello), ['Hello', 'Hello']);
+      expect(funcTimes(3, sayHello()), ['Hello', 'Hello', 'Hello']);
+      expect(funcTimes(4, () => sayHello('!')), ['Hello!', 'Hello!', 'Hello!', 'Hello!']);
+    });
   });
 }