+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add `urlJoin` method
 - Add `arrWithDefault` method
 - Add `arrWithNumber` method
+- Add `funcTimes` method
 
 ## 0.0.2 (2024-09-10)